refactoring so form parsing doesn't clutter grid routes
